requirements.txt
Django==4.2.6
mysql-connector-python==8.1.0
# mysqlclient==2.2.0
pymysql==1.1.0
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
Pillow==10.1.0
python-dateutil==2.8.2
numerize==0.12
django-summernote==0.8.20.0
beautifulsoup4==4.12.2
html5lib==1.1
pytz==2023.3
djangorestframework==3.14.0
serializers==0.2.4
python-dotenv==1.0.0
# django-ses==3.5.0
django-mail-templated==2.6.5
# redis==4.5.1
# django-redis==5.2.0
six==1.16.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.28.68
django-storages==1.14.2
django-debug-toolbar==4.2.0
bleach==6.1.0
pylibmc==1.6.3
django-hitcount==1.3.5
django-colorfield==0.10.1
requests==2.31.0